+++ src/components/ItemListContainer/ItemListContainer.js
@@ -1,0 +1,11 @@
+import './ItemListContainer.scss';
+
+const ItemListContainer = (props) => {
+    return (
+        <p>
+            {props.greeting}
+        </p>
+    );
+};
+
+export default ItemListContainer;